requirements.txt
dashscope>=1.13.0
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
openpyxl>=3.1.0
Pillow>=10.0.0
reportlab>=4.0.0
streamlit-pdf-viewer>=0.1.0
